+++ src/backends/uvc-wmf.cpp
@@ -665,7 +665,7 @@ namespace rsimpl
                     devices.push_back(dev);
                 }
 
-                int subdevice_index = mi/2;
+                size_t subdevice_index = mi/2;
                 if(subdevice_index >= dev->subdevices.size()) dev->subdevices.resize(subdevice_index+1);
 
                 dev->subdevices[subdevice_index].reader_callback = new reader_callback(dev, subdevice_index);
